+++ backend/user-service/src/routes/auth-routes.js
@@ -1,0 +1,13 @@
+const express = require("express");
+const router = express.Router();
+const {
+  registerUser,
+  verifyUser,
+  resendVerificationCode,
+} = require("../controllers/auth-controllers");
+
+router.post("/register", registerUser);
+router.post("/verify", verifyUser);
+router.post("/resend", resendVerificationCode);
+
+module.exports = router;
